+++ src/hooks/index.ts
@@ -4,3 +4,4 @@ export { useLocalStorage } from './useLocalStorage';
 export { useSessionStorage } from './useSessionStorage';
 export { useCookie } from './useCookie';
 export { useClickOutside } from './useClickOutside';
+export { usePrevious } from './usePrevious';